pyproject.toml
[project]
name = "crypto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
